+++ src/OpenTelemetry.Exporter.OneCollector/Internal/Serialization/CommonSchemaJsonSerializationState.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 // </copyright>
 
+using System.Diagnostics;
 #if NET6_0_OR_GREATER
 using System.Runtime.InteropServices;
 #endif
@@ -26,7 +27,7 @@ internal sealed class CommonSchemaJsonSerializationState
     public const int MaxNumberOfExtensionKeys = 64;
     public const int MaxNumberOfExtensionValuesPerKey = 16;
     private readonly Dictionary<string, int> keys = new(4, StringComparer.OrdinalIgnoreCase);
-    private readonly List<KeyValuePair<string, object?>> allValues = new(16);
+    private readonly List<KeyValuePair<ExtensionFieldInformation, object?>> allValues = new(16);
     private string itemType;
     private int nextKeysToAllValuesLookupIndex;
     private KeyValueLookup[] keysToAllValuesLookup = new KeyValueLookup[4];
@@ -47,11 +48,13 @@ internal sealed class CommonSchemaJsonSerializationState
     {
         if (!ExtensionFieldInformationManager.SharedCache.TryResolveExtensionFieldInformation(
             attribute.Key,
-            out (string ExtensionName, string FieldName) fieldInformation))
+            out var fieldInformation))
         {
             OneCollectorExporterEventSource.Log.AttributeDropped(this.itemType, attribute.Key, "Invalid extension field name");
             return;
         }
+
+        Debug.Assert(fieldInformation?.ExtensionName != null, "fieldInformation.ExtensionName was null");
 
 #if NET6_0_OR_GREATER
         ref var lookupIndex = ref CollectionsMarshal.GetValueRefOrAddDefault(this.keys, fieldInformation.ExtensionName, out var existed);
@@ -60,10 +63,10 @@ internal sealed class CommonSchemaJsonSerializationState
             this.AssignNewExtensionToLookupIndex(ref lookupIndex);
         }
 #else
-        if (!this.keys.TryGetValue(fieldInformation.ExtensionName, out int lookupIndex))
+        if (!this.keys.TryGetValue(fieldInformation!.ExtensionName!, out int lookupIndex))
         {
             this.AssignNewExtensionToLookupIndex(ref lookupIndex);
-            this.keys[fieldInformation.ExtensionName] = lookupIndex;
+            this.keys[fieldInformation.ExtensionName!] = lookupIndex;
         }
 #endif
 
@@ -75,14 +78,33 @@ internal sealed class CommonSchemaJsonSerializationState
 
         ref KeyValueLookup keyLookup = ref this.keysToAllValuesLookup[lookupIndex];
 
-        if (keyLookup.Count >= MaxNumberOfExtensionValuesPerKey)
+        var keyCount = keyLookup.Count;
+        if (keyCount > 0)
         {
-            OneCollectorExporterEventSource.Log.AttributeDropped(this.itemType, attribute.Key, "Extension value limit reached");
-            return;
+            if (keyCount >= MaxNumberOfExtensionValuesPerKey)
+            {
+                OneCollectorExporterEventSource.Log.AttributeDropped(this.itemType, attribute.Key, "Extension value limit reached");
+                return;
+            }
+
+            unsafe
+            {
+#if NET6_0_OR_GREATER
+                ref var firstDefinition = ref CollectionsMarshal.AsSpan(this.allValues)[keyLookup.ValueIndicies[0]];
+#else
+                var firstDefinition = this.allValues[keyLookup.ValueIndicies[0]];
+#endif
+
+                if (fieldInformation.FieldName == null || firstDefinition.Key.FieldName == null)
+                {
+                    OneCollectorExporterEventSource.Log.AttributeDropped(this.itemType, attribute.Key, "Extension is already defined");
+                    return;
+                }
+            }
         }
 
         int index = this.allValues.Count;
-        this.allValues.Add(new KeyValuePair<string, object?>(fieldInformation.FieldName, attribute.Value));
+        this.allValues.Add(new KeyValuePair<ExtensionFieldInformation, object?>(fieldInformation, attribute.Value));
 
         unsafe
         {
@@ -107,7 +129,8 @@ internal sealed class CommonSchemaJsonSerializationState
 
         foreach (var extensionPropertyKey in this.keys)
         {
-            writer.WriteStartObject(extensionPropertyKey.Key);
+            var wroteStartObject = false;
+            var wroteProperty = false;
 
             ref KeyValueLookup keyLookup = ref this.keysToAllValuesLookup[extensionPropertyKey.Value];
 
@@ -121,11 +144,37 @@ internal sealed class CommonSchemaJsonSerializationState
                     var attribute = allValues[keyLookup.ValueIndicies[i]];
 #endif
 
-                    CommonSchemaJsonSerializationHelper.SerializeKeyValueToJson(attribute.Key, attribute.Value, writer);
+                    var fieldInformation = attribute.Key;
+
+                    if (fieldInformation.FieldName == null)
+                    {
+                        Debug.Assert(!wroteProperty, "wroteProperty");
+                        Debug.Assert(!wroteStartObject, "wroteStartObject");
+
+                        writer.WritePropertyName(fieldInformation.EncodedExtensionName);
+                        CommonSchemaJsonSerializationHelper.SerializeValueToJson(attribute.Value, writer);
+                        wroteProperty = true;
+                    }
+                    else
+                    {
+                        Debug.Assert(!wroteProperty, "wroteProperty");
+
+                        if (!wroteStartObject)
+                        {
+                            writer.WriteStartObject(fieldInformation.EncodedExtensionName);
+                            wroteStartObject = true;
+                        }
+
+                        writer.WritePropertyName(fieldInformation.EncodedFieldName);
+                        CommonSchemaJsonSerializationHelper.SerializeValueToJson(attribute.Value, writer);
+                    }
                 }
             }
 
-            writer.WriteEndObject();
+            if (wroteStartObject)
+            {
+                writer.WriteEndObject();
+            }
         }
 
         if (writeExtensionObjectEnvelope)